+++ lib/rest/studio/v1/flow/execution/executionContext.d.ts
@@ -12,7 +12,6 @@ import { SerializableClass } from '../../../../../interfaces';
 
 /**
  * @description Initialize the ExecutionContextList
- * PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
  *
  * @param version - Version of the resource
  * @param flowSid - Flow Sid.
@@ -50,7 +49,7 @@ interface ExecutionContextSolution {
 
 declare class ExecutionContextContext {
   /**
-   * Initialize the ExecutionContextContextPLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
+   * Initialize the ExecutionContextContext
    *
    * @param version - Version of the resource
    * @param flowSid - Flow Sid.
@@ -69,7 +68,7 @@ declare class ExecutionContextContext {
 
 declare class ExecutionContextInstance extends SerializableClass {
   /**
-   * Initialize the ExecutionContextContextPLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
+   * Initialize the ExecutionContextContext
    *
    * @property accountSid - Account Sid.
    * @property context - Flow state.
@@ -106,7 +105,7 @@ declare class ExecutionContextInstance extends SerializableClass {
 
 declare class ExecutionContextPage extends Page<V1, ExecutionContextPayload, ExecutionContextResource, ExecutionContextInstance> {
   /**
-   * Initialize the ExecutionContextPagePLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
+   * Initialize the ExecutionContextPage
    *
    * @param version - Version of the resource
    * @param response - Response from the API